+++ public/js/router-app.js
@@ -355,10 +355,12 @@ onready(function() {
   }
 
   // Fingerprint the user
-  setTimeout(function() {
-    var fingerprint = require('./components/fingerprint');
-    fingerprint();
-  }, 5000);
+  if (context.isLoggedIn()) {
+    setTimeout(() => {
+      var fingerprint = require('./components/fingerprint');
+      fingerprint();
+    }, 5000);
+  }
 
   // Register the service worker
   if (context.hasFeature('web-push')) {