+++ public/js/vue/thread-message-feed/components/chat-item-test.js
@@ -71,6 +71,31 @@ describe('thread-message-feed chat-item', () => {
       expect(wrapper.element).toMatchSnapshot();
       expect(scrollIntoViewMock.mock.calls[0]).toEqual(['smooth', 'center']);
     });
+    it('deleted (empty) message', () => {
+      const { wrapper } = mount(
+        ChatItem,
+        {
+          ...defaultProps,
+          message: { ...message, text: '', html: '' }
+        },
+        addRoomToStore
+      );
+      expect(wrapper.element).toMatchSnapshot();
+    });
+    it('not logged in user (no message actions)', () => {
+      const { wrapper } = mount(
+        ChatItem,
+        {
+          ...defaultProps,
+          message: { ...message, text: '', html: '' }
+        },
+        store => {
+          addRoomToStore(store);
+          store.state.isLoggedIn = false;
+        }
+      );
+      expect(wrapper.element).toMatchSnapshot();
+    });
   });
   it('focusedAt - scrolls into view', () => {
     const scrollIntoViewMock = jest.fn();