+++ test/request-api-tests/chat-api-tests.js
@@ -21,6 +21,8 @@ describe('chat-api', function() {
   });
 
   var fixture = fixtureLoader.setup({
+    oAuthClient1: {},
+    oAuthAccessTokenAnonymous: { client: 'oAuthClient1', user: null },
     user1: {
       accessToken: 'web-internal'
     },
@@ -278,5 +280,32 @@ describe('chat-api', function() {
         assert.strictEqual(body.messageId, fixture.messageBad2.id);
         assert.strictEqual(body.messageText, fixture.messageBad2.text);
       });
+  });
+
+  describe('anonymous token access', () => {
+    it('GET /v1/rooms/:roomId/chatMessages', async () => {
+      const response = await request(app)
+        .get('/v1/rooms/' + fixture.troupe1.id + '/chatMessages')
+        .set('x-access-token', fixture.oAuthAccessTokenAnonymous)
+        .expect(200);
+      const messages = response.body;
+      assert(Array.isArray(messages), 'response body needs to be an array of messages');
+      assert(messages.length > 0, 'API should return some messages for anonymous token');
+      assert.equal(
+        messages.find(m => m.unread),
+        undefined,
+        'anonymous users should have all messages read'
+      );
+    });
+
+    it('PUT /v1/rooms/:roomId/chatMessages should get denied for anonymous token', async () => {
+      await request(app)
+        .post('/v1/rooms/' + fixture.troupe1.id + '/chatMessages')
+        .send({
+          text: 'Hello there'
+        })
+        .set('x-access-token', fixture.oAuthAccessTokenAnonymous)
+        .expect(401);
+    });
   });
 });